+++ types/affinda.d.ts
@@ -2031,7 +2031,9 @@ export declare interface DataFieldCreate {
 
 /** The data point to be created for this field. If a data point with the same slug and collection already exists, it will be reused. */
 export declare interface DataFieldCreateDataPoint {
+    /** Name of the data point. */
     name: string;
+    /** A camelCase string that will be used as the key in the API response. */
     slug: string;
     description?: string;
     /** The different data types of annotations */
@@ -2052,7 +2054,9 @@ export declare interface DataFieldCreateField {
 export declare interface DataFieldDataPoint {
     /** Uniquely identify a data point. */
     identifier: string;
+    /** Name of the data point. */
     name: string;
+    /** A camelCase string that will be used as the key in the API response. */
     slug: string;
     description: string | null;
     /** The different data types of annotations */
@@ -2072,18 +2076,22 @@ export declare interface DataFieldField {
 export declare interface DataPoint {
     /** Uniquely identify a data point. */
     identifier: string;
+    /** Name of the data point. */
     name: string;
-    slug?: string;
+    /** A camelCase string that will be used as the key in the API response. */
+    slug: string;
     description?: string;
     /** The different data types of annotations */
     annotationContentType: AnnotationContentType;
-    organization?: Organization;
+    organization: Organization | null;
     /** Uniquely identify an extractor. */
     extractor: string | null;
     multiple?: boolean;
     noRect?: boolean;
-    similarTo?: string[];
+    /** If true, both the value and the label for the enums will appear in the dropdown in the validation tool. */
     displayEnumValue?: boolean;
+    /** The identifier of the parent data point if applicable. */
+    parent?: string;
     children?: DataPoint[];
 }
 
@@ -2123,7 +2131,9 @@ export declare interface DataPointChoiceUpdate {
 }
 
 export declare interface DataPointCreate {
-    name?: string;
+    /** Name of the data point. */
+    name: string;
+    /** A camelCase string that will be used as the key in the API response. */
     slug: string;
     description?: string;
     /** The different data types of annotations */
@@ -2134,12 +2144,22 @@ export declare interface DataPointCreate {
     extractor: string;
     multiple?: boolean;
     noRect?: boolean;
+    /** If true, both the value and the label for the enums will appear in the dropdown in the validation tool. */
+    displayEnumValue?: boolean;
+    /** The identifier of the parent data point if applicable. */
+    parent?: string;
 }
 
 export declare interface DataPointUpdate {
+    /** Name of the data point. */
     name?: string;
+    /** A camelCase string that will be used as the key in the API response. */
     slug?: string;
     description?: string;
+    /** If true, both the value and the label for the enums will appear in the dropdown in the validation tool. */
+    displayEnumValue?: boolean;
+    /** The identifier of the parent data point if applicable. */
+    parent?: string;
 }
 
 export declare type DateAnnotation = Annotation & {