+++ CHANGELOG.md
@@ -5,6 +5,14 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Added
+- Allow create/update data point's `parent` and `displayEnumValue` property
+
+### Changed
+- Make `slug` and `organization` required when creating data point
+
+### Removed
+- Remove data point's `similarTo` property
 
 ## [6.12.0] - 2023-06-16
 ### Added